UPDATE_move_class_KeyEnums_ValueEnums_to_enums.py
