+++ src/stam_annotator/enums.py
@@ -1,0 +1,14 @@
+from enum import Enum
+
+
+class ValueEnum(Enum):
+    author = "Author"
+    book_title = "BookTitle"
+    chapter = "Chapter"
+    quotation = "Quotation"
+    sabche = "Sabche"
+    tsawa = "Tsawa"
+
+
+class KeyEnum(Enum):
+    structure_type = "Structure Type"
